requirements.txt
fastapi==0.115.12
httpx==0.26.0
loguru==0.7.3
openai==1.107.1
pydantic==2.11.7
python-dotenv==1.1.1
sentient-agent-framework==0.3.0
typing_extensions==4.15.0
uvicorn==0.34.0
tavily-python==0.3.4
